added <Leader>yw for yank word to clipboard, and enhanced transfer_vimrc for ideavimrc

+++ transfer_vimrc.sh
@@ -17,5 +17,8 @@ else
   echo "something went wrong copying vimrc.vim to ~/.vimrc"
 fi
 cp ~/vim-config/vimrc.vim ~/.vimrc 
-cat ~/vim-config/vimrc.vim | grep -v "ttimeoutlen" | grep -v "cterm" | grep -v "background" | grep -v "wildmenu" | grep -v "expandtab" | grep -v "set sw=" | grep -v "set tabstop=" | grep -v "syntax on" | grep -v "timer_start" | grep -v "cursorline" | grep -v "cursorcolumn" > ~/.ideavimrc
+if [[ -f ~/.ideavimrc ]]; then
+  message="$message as well as an optimized version of ~/.ideavimrc"
+  cat ~/vim-config/vimrc.vim | grep -v "ttimeoutlen" | grep -v "cterm" | grep -v "background" | grep -v "wildmenu" | grep -v "expandtab" | grep -v "set sw=" | grep -v "set tabstop=" | grep -v "syntax on" | grep -v "timer_start" | grep -v "cursorline" | grep -v "cursorcolumn" > ~/.ideavimrc
+fi
 echo "$message"
